Bug 588: Verified user can't login (#590)
* Fix redux-devtools-extension, add componentDidUpdate to onfido verification handler

* Fix eslint
