+++ package.json
@@ -109,35 +109,35 @@
     "web3": "1.0.0-beta.33"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.5",
-    "@babel/core": "^7.1.6",
-    "@babel/plugin-proposal-class-properties": "^7.0.0-beta.55",
-    "@babel/plugin-proposal-decorators": "^7.1.6",
-    "@babel/plugin-proposal-do-expressions": "^7.0.0-beta.55",
-    "@babel/plugin-proposal-export-default-from": "^7.0.0-beta.55",
-    "@babel/plugin-proposal-export-namespace-from": "^7.0.0-beta.55",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0-beta.55",
-    "@babel/plugin-proposal-numeric-separator": "^7.0.0-beta.55",
-    "@babel/plugin-proposal-optional-chaining": "^7.0.0-beta.55",
-    "@babel/plugin-proposal-throw-expressions": "^7.0.0-beta.55",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0-beta.55",
-    "@babel/plugin-syntax-import-meta": "^7.0.0-beta.55",
-    "@babel/preset-env": "^7.1.6",
+    "@babel/cli": "^7.2.0",
+    "@babel/core": "^7.2.0",
+    "@babel/plugin-proposal-class-properties": "^7.2.1",
+    "@babel/plugin-proposal-decorators": "^7.2.0",
+    "@babel/plugin-proposal-do-expressions": "^7.2.0",
+    "@babel/plugin-proposal-export-default-from": "^7.2.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.2.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.2.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-syntax-import-meta": "^7.2.0",
+    "@babel/preset-env": "^7.2.0",
     "@babel/preset-react": "^7.0.0-beta",
     "@babel/preset-stage-0": "^7.0.0-beta",
-    "@babel/runtime": "^7.1.5",
+    "@babel/runtime": "^7.2.0",
     "@storybook/addon-actions": "^4.0.11",
     "@storybook/addon-knobs": "^4.0.11",
     "@storybook/addon-links": "^4.0.11",
     "@storybook/react": "^4.0.11",
-    "autoprefixer": "^9.4.0",
+    "autoprefixer": "^9.4.2",
     "babel-core": "^7.0.0-0",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.2.0",
     "babel-loader": "^8.0.0-beta.2",
     "babel-plugin-dynamic-import-node": "^2.0.0",
     "babel-plugin-lodash": "^3.2.11",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.20",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.21",
     "bootstrap-loader": "^3.0.2",
     "bootstrap-sass": "^3.3.7",
     "case-sensitive-paths-webpack-plugin": "^2.1.1",
@@ -147,7 +147,7 @@
     "eslint-config-airbnb": "^17.1.0",
     "eslint-loader": "^2.0.0",
     "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^22.1.0",
+    "eslint-plugin-jest": "^22.1.2",
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-react": "^7.11.1",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
@@ -169,7 +169,7 @@
     "terser-webpack-plugin": "^1.1.0",
     "truffle-hdwallet-provider": "0.0.6",
     "url-loader": "^1.1.1",
-    "webpack": "^4.26.1",
+    "webpack": "^4.27.0",
     "webpack-cli": "^3.0.4",
     "webpack-dev-server": "^3.1.4"
   }
