+++ src/verification/onfido/components/index.js
@@ -1,6 +1,6 @@
-import React from 'react'
-import classnames from 'classnames/bind'
+import React, { Component } from 'react'
 import PropTypes from 'prop-types'
+import classnames from 'classnames/bind'
 import Indefinite from 'components/Spinner/Indefinite'
 import { getFeatureConfig } from 'utils/features'
 
@@ -33,7 +33,11 @@ const DeniedMessage = ({
     <h2>{heading || t('verification.headings.denied_default', { applicationName })}</h2>
     <p>{reason}</p>
 
-    {canRetry && <button type="button" className={cx('retryButton')} onClick={() => setStep({ page: 'welcome' })}>{t('verification.try_again')}</button>}
+    {canRetry && (
+      <button type="button" className={cx('retryButton')} onClick={() => setStep({ page: 'welcome' })}>
+        {t('verification.try_again')}
+      </button>
+    )}
   </div>
 )
 
@@ -59,15 +63,72 @@ const Steps = {
   denied: DeniedMessage,
 }
 
-const OnFido = ({ step, ...props }) => {
-  const { page, options, ...other } = step
-
-  if (Steps[page]) {
-    const StepComponent = Steps[page]
-    return <StepComponent {...other} {...props} {...options} />
+class OnFido extends Component {
+  async componentDidMount() {
+    await this.navigateUser()
   }
 
-  return <Indefinite width={64} height={64} />
+  async componentDidUpdate(prevProps) {
+    const { account } = this.props
+    if (account !== prevProps.account) {
+      await this.navigateUser()
+    }
+  }
+
+  navigateUser = async () => {
+    const {
+      tosAccepted, t, closeModal, setStep, updateUserVerification, account,
+    } = this.props
+    const applicantStatus = await updateUserVerification(account)
+
+    if (tosAccepted) {
+      if (applicantStatus === 'PENDING_DOCUMENT_UPLOAD') {
+        // TOS accepted or not required and process started but not finished - need to restart in order to get a new token
+        return setStep({
+          page: 'denied',
+          heading: t('verification.headings.application_not_completed'),
+          reason: t('verification.reasons.previous_not_completed'),
+        })
+      }
+      if (applicantStatus === 'WAITING_FOR_APPROVAL') {
+        return setStep({
+          page: 'denied',
+          heading: t('verification.headings.waiting_for_approval'),
+          reason: t('verification.reasons.verification_pending'),
+        })
+      }
+      if (applicantStatus === 'DENIED') {
+        // TOS accepted or not required and application denied
+        return setStep({
+          page: 'denied',
+          reason: t('verification.reasons.application_denied'),
+        })
+      }
+      if (applicantStatus === 'ACCEPTED') {
+        // TOS accepted and application accepted - should be able to connect now
+        // updateUserVerification will set the state entry about a finished verification if it wasnt set before.
+        return closeModal()
+      }
+    }
+
+    // TOS not accepted but required, wait until signature and message signature to let the user know if their application was denied or accepted.
+    return setStep({
+      page: 'welcome',
+    })
+  }
+
+  render() {
+    const { step, ...props } = this.props
+
+    const { page, options, ...other } = step
+
+    if (Steps[page]) {
+      const StepComponent = Steps[page]
+      return <StepComponent {...other} {...props} {...options} />
+    }
+
+    return <Indefinite width={64} height={64} />
+  }
 }
 
 OnFido.propTypes = {
