+++ src/verification/onfido/containers/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 import { reduxForm } from 'redux-form'
 import { withNamespaces } from 'react-i18next'
-import { compose, withState, lifecycle } from 'recompose'
+import { compose, withState } from 'recompose'
 
 import actions from './store/actions'
 import selectors from './store/selectors'
@@ -73,49 +73,6 @@ const enhancer = compose(
     selectors,
     actions,
   ),
-  lifecycle({
-    async componentDidMount() {
-      const {
-        updateUserVerification, setStep, account, tosAccepted, t,
-      } = this.props
-
-      const applicantStatus = await updateUserVerification(account)
-      if (tosAccepted) {
-        if (applicantStatus === 'PENDING_DOCUMENT_UPLOAD') {
-          // TOS accepted or not required and process started but not finished - need to restart in order to get a new token
-          return setStep({
-            page: 'denied',
-            heading: t('verification.headings.application_not_completed'),
-            reason: t('verification.reasons.previous_not_completed'),
-          })
-        }
-        if (applicantStatus === 'WAITING_FOR_APPROVAL') {
-          return setStep({
-            page: 'denied',
-            heading: t('verification.headings.waiting_for_approval'),
-            reason: t('verification.reasons.verification_pending'),
-          })
-        }
-        if (applicantStatus === 'DENIED') {
-          // TOS accepted or not required and application denied
-          return setStep({
-            page: 'denied',
-            reason: t('verification.reasons.application_denied'),
-          })
-        }
-        if (applicantStatus === 'ACCEPTED') {
-          // TOS accepted and application accepted - should be able to connect now
-          // updateUserVerification will set the state entry about a finished verification if it wasnt set before.
-          return this.props.closeModal()
-        }
-      }
-
-      // TOS not accepted but required, wait until signature and message signature to let the user know if their application was denied or accepted.
-      return setStep({
-        page: 'welcome',
-      })
-    },
-  }),
   reduxForm({
     form: 'VERIFICATION_ONFIDO',
     onSubmit: (fields) => {
